pyproject.toml
[tool.poetry]
name = "evadata"
version = "0.1.0"
description = ""
authors = ["ICheered <tjeerd992@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.23.3"
netCDF4 = "^1.6.1"
Pillow = "^9.2.0"
sklearn = "^0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
